pyproject.toml
[project]
name = "swim"
version = "0.1.0"
description = "Campus Software Image Management (SWIM)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "django-auth-ldap>=4.6.0",
    "django-cors-headers>=4.3.0",
    "pyats[full]>=24.1",
    "genie>=24.1",
    "scp>=0.14.5",
    "paramiko>=3.4.0",
    "python-dotenv>=1.0.0"
]
